# templates/service/api/openapi.yaml
...
openapi: "3.0.3"
info:
  title: __SERVICE_FULL__ API
  description: |
    The __SERVICE_FULL__ service manages __RESOURCE_LOWER__ resources for the
    OpenCHAMI system management platform.
  version: "__API_VERSION__"
  contact:
    name: OpenCHAMI
    url: https://git.cscs.ch/openchami/__SERVICE_FULL__
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:__PORT__
    description: Local development
...
paths:
...
  __API_PREFIX__/__RESOURCE_PLURAL__:
    get:
      tags: [__RESOURCE_PLURAL__]
      summary: List __RESOURCE_LOWER__ resources
      operationId: list__RESOURCE__s
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        # TEMPLATE: Add filter parameters here.
        # - name: type
        #   in: query
        #   schema:
        #     type: string
        #   description: Filter by resource type
      responses:
        "200":
          description: Paginated list of __RESOURCE_LOWER__ resources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/__RESOURCE__List"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [__RESOURCE_PLURAL__]
      summary: Create a new __RESOURCE_LOWER__
      operationId: create__RESOURCE__
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Create__RESOURCE__Request"
      responses:
        "201":
          description: __RESOURCE__ created
          headers:
            Location:
              description: URI of the created resource
              schema:
                type: string
                example: "__API_PREFIX__/__RESOURCE_PLURAL__/550e8400-e29b-41d4-a716-446655440000"
            ETag:
              description: Entity tag for the created resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/__RESOURCE__Resource"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"

  # =========================================================================
  # __RESOURCE__ Single
  # =========================================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return (1-200, default 50)
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      name: offset
      in: query
      description: Number of items to skip for pagination
      schema:
        type: integer
        minimum: 0
        default: 0

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    __RESOURCE__Resource:
      type: object
      required: [kind, apiVersion, metadata, spec]
      properties:
        kind:
          type: string
          example: "__RESOURCE__"
        apiVersion:
          type: string
          example: "__API_VERSION__"
        metadata:
          $ref: "#/components/schemas/ResourceMetadata"
        spec:
          $ref: "#/components/schemas/__RESOURCE__"

    # -- Resource envelope (list) ------------------------------------------
    __RESOURCE__List:
      type: object
      required: [kind, apiVersion, metadata, items]
      properties:
        kind:
          type: string
          example: "__RESOURCE__List"
        apiVersion:
          type: string
          example: "__API_VERSION__"
        metadata:
          $ref: "#/components/schemas/ListMetadata"
        items:
          type: array
          items:
            $ref: "#/components/schemas/__RESOURCE__"

    # -- Metadata ----------------------------------------------------------
    ResourceMetadata:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ListMetadata:
      type: object
      properties:
        totalCount:
          type: integer
          example: 42
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0

    # -- __RESOURCE__ spec --------------------------------------------------
    # TEMPLATE: Define all fields of your resource here.
    __RESOURCE__:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Human-readable name
          example: "my-__RESOURCE_LOWER__"
        description:
          type: string
          description: Optional description
          example: "A sample __RESOURCE_LOWER__ resource"
        # TEMPLATE: Add resource-specific fields.
        # type:
        #   type: string
        #   enum: [TypeA, TypeB]
        # status:
        #   type: string
        #   enum: [active, inactive]

    # -- Request bodies ----------------------------------------------------
    Create__RESOURCE__Request:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "my-__RESOURCE_LOWER__"
        description:
          type: string
          example: "A new __RESOURCE_LOWER__"
...
    ProblemDetail:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          example: "about:blank"
        title:
          type: string
          example: "Not Found"
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: '__RESOURCE_LOWER__ "abc123" not found'
        instance:
          type: string
          example: "__API_PREFIX__/__RESOURCE_PLURAL__/abc123"
        errors:
          type: array
          description: Field-level validation errors (present on 400 responses)
          items:
            type: object
            properties:
              field:
                type: string
                example: "name"
              message:
                type: string
                example: "name is required and must not be blank"

  # -------------------------------------------------------------------------
  # Reusable Responses
  # -------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid request (validation error)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    Conflict:
      description: Resource conflict (duplicate)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
security:
  - BearerAuth: []